Initializes the pip core packages

Based on options() settings provides an option to download latest package versions from the branch

Usage

init_metapip(exclude = NA, ask = TRUE, answer = 1)

update_pip_packages(exclude = NA, ask = TRUE, answer = 1)

Arguments

exclude

character: packages to exclude from attaching. if [getwd] is one of the core PIP packages, that package will be excluded be default. To avoid that, set exclude to `NULL`.

ask

logical. Ask the user if she wants to install outdated packages. Default TRUE

answer

numeric: Developers argument. Only works for demonstration purposes.

Value

`init_metapip()` return invisible() output

`update_pip_packages()` return logical vector. TRUE if missing package were update. FALSE if all packages are up to date of the user selects not to update.

Examples

if (FALSE) { # \dontrun{
  init_metapip()
} # }

update_pip_packages(ask = FALSE,
answer = 2) # this is to make it work in examples and vignettes.
#> ! The following packages do not have the updated version of default branch
#> installed: pipapi, pipload, wbpip, pipfun, pipdata, pipster, pipaux, and pipfaker
#> ℹ Installing missing packages...
#> ℹ Installing branch DEV from package pipapi
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Downloading GitHub repo PIP-Technical-Team/pipapi@DEV
#> wbpip        (NA -> fd2c687ed...) [GitHub]
#> bit          (NA -> 4.6.0       ) [CRAN]
#> cpp11        (NA -> 0.5.2       ) [CRAN]
#> prettyunits  (NA -> 1.2.0       ) [CRAN]
#> progress     (NA -> 1.2.3       ) [CRAN]
#> tzdb         (NA -> 0.5.0       ) [CRAN]
#> tidyselect   (NA -> 1.2.1       ) [CRAN]
#> hms          (NA -> 1.1.3       ) [CRAN]
#> bit64        (NA -> 4.6.0-1     ) [CRAN]
#> RcppParallel (NA -> 5.1.11-1    ) [CRAN]
#> later        (NA -> 1.4.4       ) [CRAN]
#> promises     (NA -> 1.3.3       ) [CRAN]
#> DBI          (NA -> 1.2.3       ) [CRAN]
#> data.table   (NA -> 1.17.8      ) [CRAN]
#> assertthat   (NA -> 0.2.1       ) [CRAN]
#> vroom        (NA -> 1.6.5       ) [CRAN]
#> clipr        (NA -> 0.8.0       ) [CRAN]
#> BH           (NA -> 1.87.0-1    ) [CRAN]
#> stringfish   (NA -> 0.17.0      ) [CRAN]
#> RApiSeria... (NA -> 0.1.4       ) [CRAN]
#> triebeard    (NA -> 0.4.1       ) [CRAN]
#> swagger      (NA -> 5.17.14.1   ) [CRAN]
#> sodium       (NA -> 1.4.0       ) [CRAN]
#> httpuv       (NA -> 1.6.16      ) [CRAN]
#> webutils     (NA -> 1.2.2       ) [CRAN]
#> fstcore      (NA -> 0.10.0      ) [CRAN]
#> duckdb       (NA -> 1.4.0       ) [CRAN]
#> joyn         (NA -> 0.2.4       ) [CRAN]
#> arrow        (NA -> 21.0.0.1    ) [CRAN]
#> readr        (NA -> 2.1.5       ) [CRAN]
#> qs           (NA -> 0.27.3      ) [CRAN]
#> urltools     (NA -> 1.7.3.1     ) [CRAN]
#> plumber      (NA -> 1.3.0       ) [CRAN]
#> fst          (NA -> 0.9.8       ) [CRAN]
#> Downloading GitHub repo PIP-Technical-Team/wbpip@DEV
#> tidyselect (NA -> 1.2.1 ) [CRAN]
#> generics   (NA -> 0.1.4 ) [CRAN]
#> assertthat (NA -> 0.2.1 ) [CRAN]
#> data.table (NA -> 1.17.8) [CRAN]
#> dplyr      (NA -> 1.1.4 ) [CRAN]
#> Installing 5 packages: tidyselect, generics, assertthat, data.table, dplyr
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d198eedcb9/PIP-Technical-Team-wbpip-fd2c687/DESCRIPTION’ ... OK
#> * preparing ‘wbpip’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#>   NB: this package now depends on R (>= 4.1.0)
#>   WARNING: Added dependency on R >= 4.1.0 because package code uses the
#>   pipe |> or function shorthand \(...) syntax added in R 4.1.0.
#>   File(s) using such syntax:
#>     ‘gd_compute_pip_stats_lb.R’ ‘gd_compute_pip_stats_lq.R’
#>     ‘low-hanging_indicators.R’ ‘md_compute_polarization.R’
#>     ‘md_compute_poverty_stats.R’ ‘md_infer_poverty_line.R’
#> * building ‘wbpip_0.1.6.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Installing 33 packages: bit, cpp11, prettyunits, progress, tzdb, tidyselect, hms, bit64, RcppParallel, later, promises, DBI, data.table, assertthat, vroom, clipr, BH, stringfish, RApiSerialize, triebeard, swagger, sodium, httpuv, webutils, fstcore, duckdb, joyn, arrow, readr, qs, urltools, plumber, fst
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Skipping install of 'wbpip' from a github remote, the SHA1 (fd2c687e) has not changed since last install.
#>   Use `force = TRUE` to force installation
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d191c6d5a39/PIP-Technical-Team-pipapi-c001558/DESCRIPTION’ ... OK
#> * preparing ‘pipapi’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#>   NB: this package now depends on R (>= 4.3.0)
#>   WARNING: Added dependency on R >= 4.3.0 because package code uses the
#>   pipe placeholder at the head of a chain of extractions syntax added
#>   in R 4.3.0.
#>   File(s) using such syntax:
#>     ‘utils.R’
#> * building ‘pipapi_1.3.23.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ℹ Installing branch DEV_v2 from package pipload
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Downloading GitHub repo PIP-Technical-Team/pipload@DEV_v2
#> pipfun      (NA -> cbb01b10f...) [GitHub]
#> zip         (NA -> 2.3.3       ) [CRAN]
#> credentials (NA -> 2.0.3       ) [CRAN]
#> httr        (NA -> 1.4.7       ) [CRAN]
#> gert        (NA -> 2.1.5       ) [CRAN]
#> forcats     (NA -> 1.0.0       ) [CRAN]
#> pins        (NA -> 1.4.1       ) [CRAN]
#> usethis     (NA -> 3.2.1       ) [CRAN]
#> haven       (NA -> 2.5.5       ) [CRAN]
#> Downloading GitHub repo PIP-Technical-Team/pipfun@DEV_v2
#> rematch    (NA -> 2.0.0) [CRAN]
#> cellranger (NA -> 1.1.0) [CRAN]
#> forcats    (NA -> 1.0.0) [CRAN]
#> readxl     (NA -> 1.4.5) [CRAN]
#> haven      (NA -> 2.5.5) [CRAN]
#> Installing 5 packages: rematch, cellranger, forcats, readxl, haven
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d191978512f/PIP-Technical-Team-pipfun-cbb01b1/DESCRIPTION’ ... OK
#> * preparing ‘pipfun’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * looking to see if a ‘data/datalist’ file should be added
#>   NB: this package now depends on R (>= 4.3.0)
#>   WARNING: Added dependency on R >= 4.3.0 because package code uses the
#>   pipe placeholder at the head of a chain of extractions syntax added
#>   in R 4.3.0.
#>   File(s) using such syntax:
#>     ‘working_release.R’
#> * building ‘pipfun_0.3.11.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Installing 8 packages: zip, credentials, httr, gert, forcats, pins, usethis, haven
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Skipping install of 'pipfun' from a github remote, the SHA1 (cbb01b10) has not changed since last install.
#>   Use `force = TRUE` to force installation
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d1957607b7b/PIP-Technical-Team-pipload-b2e7571/DESCRIPTION’ ... OK
#> * preparing ‘pipload’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> Omitted ‘LazyData’ from DESCRIPTION
#>   NB: this package now depends on R (>= 4.1.0)
#>   WARNING: Added dependency on R >= 4.1.0 because package code uses the
#>   pipe |> or function shorthand \(...) syntax added in R 4.1.0.
#>   File(s) using such syntax:
#>     ‘load_dlw_data.R’ ‘load_pip_data.R’ ‘pip_load_all_aux.R’
#>     ‘pip_load_aux.R’ ‘pip_load_cache.R’ ‘pip_load_cache_inventory.R’
#>     ‘pip_merge_aux.R’ ‘pip_read-write.R’ ‘pip_update_inventory.R’
#>     ‘piploadenv.R’ ‘utils.R’
#> * building ‘pipload_0.2.4.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ℹ Installing branch DEV from package wbpip
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Skipping install of 'wbpip' from a github remote, the SHA1 (fd2c687e) has not changed since last install.
#>   Use `force = TRUE` to force installation
#> ℹ Installing branch DEV_v2 from package pipfun
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Skipping install of 'pipfun' from a github remote, the SHA1 (cbb01b10) has not changed since last install.
#>   Use `force = TRUE` to force installation
#> ℹ Installing branch DEV_v2 from package pipdata
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Downloading GitHub repo PIP-Technical-Team/pipdata@DEV_v2
#> wbpip   (fd2c687ed... -> NA          ) [GitHub]
#> pipload (b2e7571f6... -> NA          ) [GitHub]
#> pipfun  (cbb01b10f... -> 3d8092aa5...) [GitHub]
#> Downloading GitHub repo pip-technical-team/wbpip@dev
#> 
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d193c1fb34f/PIP-Technical-Team-wbpip-fd2c687/DESCRIPTION’ ... OK
#> * preparing ‘wbpip’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#>   NB: this package now depends on R (>= 4.1.0)
#>   WARNING: Added dependency on R >= 4.1.0 because package code uses the
#>   pipe |> or function shorthand \(...) syntax added in R 4.1.0.
#>   File(s) using such syntax:
#>     ‘gd_compute_pip_stats_lb.R’ ‘gd_compute_pip_stats_lq.R’
#>     ‘low-hanging_indicators.R’ ‘md_compute_polarization.R’
#>     ‘md_compute_poverty_stats.R’ ‘md_infer_poverty_line.R’
#> * building ‘wbpip_0.1.6.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Downloading GitHub repo PIP-Technical-Team/pipload@new_pipeline
#> Error in utils::download.file(url, path, method = method, quiet = quiet,  : 
#>   cannot open URL 'https://api.github.com/repos/PIP-Technical-Team/pipload/tarball/new_pipeline'
#> Downloading GitHub repo PIP-Technical-team/pipfun@ongoing
#> 
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d197c4d4009/PIP-Technical-Team-pipfun-3d8092a/DESCRIPTION’ ... OK
#> * preparing ‘pipfun’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#>   NB: this package now depends on R (>= 4.1.0)
#>   WARNING: Added dependency on R >= 4.1.0 because package code uses the
#>   pipe |> or function shorthand \(...) syntax added in R 4.1.0.
#>   File(s) using such syntax:
#>     ‘load_from_gh.R’ ‘pip_create_globals.R’ ‘pip_merge_aux.R’
#> * building ‘pipfun_0.0.0.9018.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d19340e631d/PIP-Technical-Team-pipdata-fb44874/DESCRIPTION’ ... OK
#> * preparing ‘pipdata’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> Removed empty directory ‘pipdata/vignettes’
#> * looking to see if a ‘data/datalist’ file should be added
#>   NB: this package now depends on R (>= 4.1.0)
#>   WARNING: Added dependency on R >= 4.1.0 because package code uses the
#>   pipe |> or function shorthand \(...) syntax added in R 4.1.0.
#>   File(s) using such syntax:
#>     ‘pd_add_pip_vars.R’ ‘pd_dlw_clean.R’ ‘update_dlw_inventory.R’
#>     ‘utils.R’
#> * building ‘pipdata_0.0.0.9017.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ℹ Installing branch DEV from package pipster
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Downloading GitHub repo PIP-Technical-Team/pipster@DEV
#> wbpip (fd2c687ed... -> NA) [GitHub]
#> Downloading GitHub repo PIP-Technical-Team/wbpip@dev_stage2
#> Error in utils::download.file(url, path, method = method, quiet = quiet,  : 
#>   cannot open URL 'https://api.github.com/repos/PIP-Technical-Team/wbpip/tarball/dev_stage2'
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d191e5c8ef6/PIP-Technical-Team-pipster-1f6b1c8/DESCRIPTION’ ... OK
#> * preparing ‘pipster’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#>   NB: this package now depends on R (>= 3.5.0)
#>   WARNING: Added dependency on R >= 3.5.0 because serialized objects in
#>   serialize/load version 3 cannot be read in older versions of R.
#>   File(s) containing such objects:
#>     ‘pipster/tests/testthat/testdata/gd_ex1.RDS’
#>     ‘pipster/tests/testthat/testdata/gd_ex2.RDS’
#>     ‘pipster/tests/testthat/testdata/local/ago2008.RDS’
#>     ‘pipster/tests/testthat/testdata/local/ago2018.RDS’
#>     ‘pipster/tests/testthat/testdata/local/bdi2013.RDS’
#>     ‘pipster/tests/testthat/testdata/local/idn1990.RDS’
#>     ‘pipster/tests/testthat/testdata/local/nga1996.RDS’
#>     ‘pipster/tests/testthat/testdata/local/nga2003.RDS’
#>     ‘pipster/tests/testthat/testdata/local/ukr1995.RDS’
#>     ‘pipster/tests/testthat/testdata/local/ukr1999.RDS’
#>     ‘pipster/tests/testthat/testdata/local/ukr2002.RDS’
#>     ‘pipster/tests/testthat/testdata/local/zwe2011.RDS’
#>     ‘pipster/tests/testthat/testdata/local/zwe2017.RDS’
#>     ‘pipster/tests/testthat/testdata/md_lorenz1.RDS’
#>     ‘pipster/tests/testthat/testdata/md_lorenz2.RDS’
#>     ‘pipster/tests/testthat/testdata/md_lorenz3.RDS’
#>     ‘pipster/tests/testthat/testdata/synthetic-microdata.RDS’
#>   NB: this package now depends on R (>= 4.2.0)
#>   WARNING: Added dependency on R >= 4.2.0 because package code uses the
#>   pipe placeholder syntax added in R 4.2.0
#>   File(s) using such syntax:
#>     ‘create_pipster_object.R’ ‘pipgd_dist.R’ ‘pipgd_pov.R’
#> * building ‘pipster_0.0.1.9003.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ℹ Installing branch DEV_v2 from package pipaux
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Downloading GitHub repo PIP-Technical-Team/pipaux@DEV_v2
#> pipfun       (3d8092aa5... -> cbb01b10f...) [GitHub]
#> commonmark   (NA           -> 2.0.0       ) [CRAN]
#> sourcetools  (NA           -> 0.1.7-1     ) [CRAN]
#> xtable       (NA           -> 1.8-4       ) [CRAN]
#> shiny        (NA           -> 1.11.1      ) [CRAN]
#> semantic.... (NA           -> 1.1.0       ) [CRAN]
#> htmlwidgets  (NA           -> 1.6.4       ) [CRAN]
#> timechange   (NA           -> 0.3.0       ) [CRAN]
#> uuid         (NA           -> 1.2-1       ) [CRAN]
#> here         (NA           -> 1.0.2       ) [CRAN]
#> getPass      (NA           -> 0.2-4       ) [CRAN]
#> tidyr        (NA           -> 1.3.1       ) [CRAN]
#> shiny.sem... (NA           -> 0.5.1       ) [CRAN]
#> assertr      (NA           -> 3.0.1       ) [CRAN]
#> lubridate    (NA           -> 1.9.4       ) [CRAN]
#> blastula     (NA           -> 0.3.6       ) [CRAN]
#> data.vali... (NA           -> 0.2.1       ) [CRAN]
#> wbstats      (NA           -> 1.1         ) [CRAN]
#> Installing 17 packages: commonmark, sourcetools, xtable, shiny, semantic.assets, htmlwidgets, timechange, uuid, here, getPass, tidyr, shiny.semantic, assertr, lubridate, blastula, data.validator, wbstats
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Downloading GitHub repo PIP-Technical-Team/pipfun@DEV_v2
#> 
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d192219dbda/PIP-Technical-Team-pipfun-cbb01b1/DESCRIPTION’ ... OK
#> * preparing ‘pipfun’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * looking to see if a ‘data/datalist’ file should be added
#>   NB: this package now depends on R (>= 4.3.0)
#>   WARNING: Added dependency on R >= 4.3.0 because package code uses the
#>   pipe placeholder at the head of a chain of extractions syntax added
#>   in R 4.3.0.
#>   File(s) using such syntax:
#>     ‘working_release.R’
#> * building ‘pipfun_0.3.11.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d1934c9ef9e/PIP-Technical-Team-pipaux-c8bc347/DESCRIPTION’ ... OK
#> * preparing ‘pipaux’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> Omitted ‘LazyData’ from DESCRIPTION
#>   NB: this package now depends on R (>= 4.2.0)
#>   WARNING: Added dependency on R >= 4.2.0 because package code uses the
#>   pipe placeholder syntax added in R 4.2.0
#>   File(s) using such syntax:
#>     ‘aux_fun.R’ ‘identify_changes.R’ ‘old_update_aux.R’
#> * building ‘pipaux_0.3.0.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ℹ Installing branch main from package pipfaker
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Downloading GitHub repo PIP-Technical-Team/pipfaker@main
#> wbpip (fd2c687ed... -> NA) [GitHub]
#> Downloading GitHub repo pip-technical-team/wbpip@dev
#> 
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d196c1f7490/PIP-Technical-Team-wbpip-fd2c687/DESCRIPTION’ ... OK
#> * preparing ‘wbpip’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#>   NB: this package now depends on R (>= 4.1.0)
#>   WARNING: Added dependency on R >= 4.1.0 because package code uses the
#>   pipe |> or function shorthand \(...) syntax added in R 4.1.0.
#>   File(s) using such syntax:
#>     ‘gd_compute_pip_stats_lb.R’ ‘gd_compute_pip_stats_lq.R’
#>     ‘low-hanging_indicators.R’ ‘md_compute_polarization.R’
#>     ‘md_compute_poverty_stats.R’ ‘md_infer_poverty_line.R’
#> * building ‘wbpip_0.1.6.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpJNXEMK/remotes1d19309313d2/PIP-Technical-Team-pipfaker-5070978/DESCRIPTION’ ... OK
#> * preparing ‘pipfaker’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#>   NB: this package now depends on R (>= 4.1.0)
#>   WARNING: Added dependency on R >= 4.1.0 because package code uses the
#>   pipe |> or function shorthand \(...) syntax added in R 4.1.0.
#>   File(s) using such syntax:
#>     ‘fk_micro_gen.R’ ‘fk_pip.R’ ‘utils.R’
#> * building ‘pipfaker_0.2.1.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)